requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.2
django-cleanup==9.0.0
django-cors-headers==4.5.0
djangorestframework==3.15.2
pillow==11.0.0
psycopg2==2.9.10
pygame==2.1.3.dev8
python-barcode==0.15.1
python-nmap==0.7.1
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
